refactor(workshop): move handlers into feature module
