+++ packages/core/src/__tests__/application/handlers/domain/WorkshopHandler.test.ts
@@ -2,7 +2,7 @@ import {
   isWorkshopHostReturnShortcut,
   WorkshopHandler,
   WorkshopWidgetRuntime
-} from '@/application/handlers/domain/WorkshopHandler';
+} from '@/application/handlers/domain/workshop/WorkshopHandler';
 import { WorkshopSessionService } from '@/application/services/workshop/WorkshopSessionService';
 import {
   WorkshopRoomDeliveryService
