+++ packages/core/src/application/handlers/domain/workshop/widgets/lexicalGravity/WorkshopLexicalGravityHandler.ts
@@ -2,7 +2,7 @@
 
 import { MessageRouter } from '@/application/handlers/MessageRouter';
 import { MessageTransport } from '@/application/handlers/MessageHandlerContracts';
-import { WorkshopMutationRouteRegistrar } from '@handlers/domain/WorkshopSessionMessageHandler';
+import { WorkshopMutationRouteRegistrar } from '@handlers/domain/workshop/WorkshopSessionMessageHandler';
 import { LexicalGravityLensRepository } from '@/infrastructure/storage/LexicalGravityLensRepository';
 import { LexicalGravityModelService } from '@services/widgets/LexicalGravityModelService';
 import {
